Created basic image add in UI
Added in the ImageClicked command to check whether the sender (clicked image) is the floor image. If it is, then a new image is added at the point of click. The functionality is intended for settings page, but due to uneven implementation rates it was more convenient to test it in the home page.