Dummy data with behavior in UI
After a lot of fiddling with the ItemsControl element I finally was able to bind an observable collection of cameras to a list of cameras of a floor. The observable collection's contents are bound to itemsControl and each item has its own image, that is located based on PositionX and PositionY, both of which belong to camera. Also now elements of ItemSource react to user interaction.