Fake data and scrolling
Whenever you click on a camera, you add a viking named randomly to the identified people's list. When not all items fit in the DataGrid, scrolling is enabled.
Also made the ViewBox stay in the top of the screen, as it used to go up and down due to window resizing as by default it was centered.